Lab 02 has been done! 2 commit

+++ sources/task5.cpp
@@ -1,13 +1,16 @@
-#include "lab05.hpp"
+//
+// Created by Ivan on 10.10.2022.
+//
+#include <lab05.hpp>
 
 bool remove_first_negative_element(std::vector<int>& vec,
                                    int& removed_element) {
-  int k = 0;
+  size_t k = 0;
   for (int i : vec) {
     k++;
     if (i < 0) {
       removed_element = i;
-      for (; k - 1 < int(vec.size()); k++) {
+      for (; size_t(k - 1) < vec.size(); k++) {
         vec[k - 1] = vec[k];
       }
       vec.resize(int(vec.size()) - 1);
